:grey_exclamation: Add `--force` to fetch command
The the Git documentation
```
When git fetch is used with <rbranch>:<lbranch> refspec, it refuses to update the local branch <lbranch> unless the remote branch <rbranch> it fetches is a descendant of <lbranch>. This option overrides that check.
```

We always want our local bare repo to reflect the remote repo.